# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Training Program Service API
  version: 0.2.0
  description: |
    Stateful API contract for the workout program service.

    Auth is local to this service for the MVP: unique nickname plus password.
    Authenticated requests use an HttpOnly `sid` cookie.

    The `init=1` cookie is a temporary JavaScript-readable bootstrap marker.
    It is not authentication and must not be used as authorization.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/auth/session:
    get:
      tags: [auth]
      operationId: getAuthSession
      summary: Probe current session without requiring authentication
      parameters:
        - $ref: "#/components/parameters/SidCookie"
      responses:
        "200":
          description: Current session state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
...
  /v1/me/profile:
    get:
      tags: [auth]
      operationId: getMyProfile
      summary: Get athlete profile defaults
      security:
        - SessionCookie: []
      responses:
        "200":
          description: Athlete profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AthleteProfileResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      tags: [auth]
      operationId: putMyProfile
      summary: Save athlete profile defaults
      description: Profile changes do not automatically rewrite existing cycle settings.
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutAthleteProfileRequest"
      responses:
        "200":
          description: Athlete profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AthleteProfileResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/program/options:
    get:
      tags: [program]
      operationId: getProgramOptions
      summary: Get selectable program options
      security:
        - SessionCookie: []
      responses:
        "200":
          description: Program options
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramOptionsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/program/calculate:
    post:
      tags: [program]
      operationId: calculateProgram
      summary: Calculate training days from settings and week
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalculateProgramRequest"
      responses:
        "200":
          description: Calculated training plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingPlanResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/cycles:
    get:
      tags: [cycles]
      operationId: listCycles
      summary: List user's cycles
      security:
        - SessionCookie: []
      responses:
        "200":
          description: Cycle history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CyclesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [cycles]
      operationId: createCycle
      summary: Create a new active cycle
      description: Creating a cycle archives the user's previous active cycle.
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutCurrentCycleRequest"
      responses:
        "201":
          description: Created active cycle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CycleResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/cycles/{cycleId}:
    put:
      tags: [cycles]
      operationId: putCycle
      summary: Update a cycle by id
      security:
        - SessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CycleIDPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutCurrentCycleRequest"
      responses:
        "200":
          description: Updated cycle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CycleResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/cycles/{cycleId}/activate:
    post:
      tags: [cycles]
      operationId: activateCycle
      summary: Make an existing cycle active
      description: Activating a cycle archives the user's previous active cycle.
      security:
        - SessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CycleIDPath"
      responses:
        "200":
          description: Activated cycle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CycleResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/cycles/current:
    get:
      tags: [cycles]
      operationId: getCurrentCycle
      summary: Get current active cycle with editable settings
      security:
        - SessionCookie: []
      responses:
        "200":
          description: Current active cycle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CycleResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [cycles]
      operationId: putCurrentCycle
      summary: Create or update current active cycle settings
      description: Settings may be changed while progressing through the cycle.
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutCurrentCycleRequest"
      responses:
        "200":
          description: Current active cycle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CycleResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/cycles/current/advance:
    post:
      tags: [cycles]
      operationId: advanceCurrentCycle
      summary: Move the current cycle to another week
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdvanceCycleRequest"
      responses:
        "200":
          description: Updated cycle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CycleResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/cycles/current/plan:
    get:
      tags: [cycles]
      operationId: getCurrentCyclePlan
      summary: Calculate the current active cycle plan
      security:
        - SessionCookie: []
      responses:
        "200":
          description: Current active cycle training plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingPlanResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/cycles/current/progress:
    get:
      tags: [cycles]
      operationId: listCurrentCycleProgress
      summary: List checkpoints for the current active cycle
      security:
        - SessionCookie: []
      parameters:
        - $ref: "#/components/parameters/WeekQuery"
      responses:
        "200":
          description: Progress checkpoints
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressCheckpointsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/cycles/current/progress/checkpoints:
    put:
      tags: [cycles]
      operationId: upsertCurrentCycleCheckpoint
      summary: Create or update one progress checkpoint
      description: Upserts by `(cycle, week, dayId, exerciseKey)` for the MVP.
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertProgressCheckpointRequest"
      responses:
        "200":
          description: Upserted checkpoint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressCheckpointResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/cycles/current/progress/checkpoints/{checkpointId}:
    delete:
      tags: [cycles]
      operationId: deleteCurrentCycleCheckpoint
      summary: Delete one progress checkpoint
      security:
        - SessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CheckpointIDPath"
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/exercises:
    get:
      tags: [exercises]
      operationId: listExercises
      summary: Search dataset exercise catalog
      security:
        - SessionCookie: []
      parameters:
        - $ref: "#/components/parameters/ExerciseQuery"
        - $ref: "#/components/parameters/LimitQuery"
        - $ref: "#/components/parameters/OffsetQuery"
        - $ref: "#/components/parameters/HasGifQuery"
      responses:
        "200":
          description: Exercise catalog results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExerciseCatalogListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/exercises/catalog/{datasetExerciseId}:
    get:
      tags: [exercises]
      operationId: getCatalogExercise
      summary: Get details for a dataset exercise
      security:
        - SessionCookie: []
      parameters:
        - $ref: "#/components/parameters/DatasetExerciseIDPath"
      responses:
        "200":
          description: Dataset exercise details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExerciseCatalogItemResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/exercises/{exerciseKey}:
    get:
      tags: [exercises]
      operationId: getExerciseDetails
      summary: Get details for a program exercise
      security:
        - SessionCookie: []
      parameters:
        - $ref: "#/components/parameters/ExerciseKeyPath"
      responses:
        "200":
          description: Exercise details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExerciseDetailsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: sid
      description: HttpOnly authenticated session cookie.
    InitCookie:
      type: apiKey
      in: cookie
      name: init
      description: Temporary bootstrap marker. Not authentication.
...
  parameters:
    WeekQuery:
      name: week
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/ProgramWeek"
    SidCookie:
      name: sid
      in: cookie
      required: false
      schema:
        type: string
    CycleIDPath:
      name: cycleId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CheckpointIDPath:
      name: checkpointId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    HasGifQuery:
      name: hasGif
      in: query
      required: false
      schema:
        type: boolean
  responses:
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    SessionResponse:
      type: object
      required: [authenticated]
      properties:
        authenticated:
          type: boolean
        user:
          $ref: "#/components/schemas/User"
...
    AthleteProfileResponse:
      type: object
      required: [profile]
      properties:
        profile:
          $ref: "#/components/schemas/AthleteProfile"
    PutAthleteProfileRequest:
      type: object
      required: [profile]
      properties:
        profile:
          $ref: "#/components/schemas/AthleteProfileInput"
...
    ProgramOptionsResponse:
      type: object
      required: [weeks, variants, progressionSteps, assistance, gpp]
      properties:
        weeks:
          type: array
          items:
            $ref: "#/components/schemas/SelectOption"
        variants:
          type: array
          items:
            $ref: "#/components/schemas/SelectOption"
        progressionSteps:
          type: array
          items:
            $ref: "#/components/schemas/SelectOption"
        assistance:
          $ref: "#/components/schemas/AssistanceOptions"
        gpp:
          $ref: "#/components/schemas/GPPOptions"
...
    CalculateProgramRequest:
      type: object
      required: [selection]
      properties:
        selection:
          $ref: "#/components/schemas/ProgramSelection"
...
    CycleResponse:
      type: object
      required: [cycle]
      properties:
        cycle:
          $ref: "#/components/schemas/ProgramCycle"
    CyclesResponse:
      type: object
      required: [cycles]
      properties:
        cycles:
          type: array
          items:
            $ref: "#/components/schemas/ProgramCycle"
        currentCycleId:
          type: string
          nullable: true
          format: uuid
...
    PutCurrentCycleRequest:
      type: object
      required: [title, currentWeek, settings]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
        currentWeek:
          $ref: "#/components/schemas/ProgramWeek"
        settings:
          $ref: "#/components/schemas/CycleSettings"
        changeReason:
          type: string
          nullable: true
          maxLength: 500
    AdvanceCycleRequest:
      type: object
      required: [targetWeek]
      properties:
        targetWeek:
          $ref: "#/components/schemas/ProgramWeek"
        note:
          type: string
          nullable: true
          maxLength: 500
    TrainingPlanResponse:
      type: object
      required: [selection, days, formulaVersion, warnings]
      properties:
        selection:
          $ref: "#/components/schemas/ProgramSelection"
        days:
          type: array
          items:
            $ref: "#/components/schemas/TrainingDay"
        formulaVersion:
          type: string
          example: xlsx-linear-cycle-v1
        warnings:
          type: array
          items:
            type: string
...
    ProgressCheckpointsResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ProgressCheckpoint"
    ProgressCheckpointResponse:
      type: object
      required: [checkpoint]
      properties:
        checkpoint:
          $ref: "#/components/schemas/ProgressCheckpoint"
    UpsertProgressCheckpointRequest:
      type: object
      required: [checkpoint]
      properties:
        checkpoint:
          $ref: "#/components/schemas/ProgressCheckpointInput"
...